backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic==2.6.4
python-dotenv==1.0.1

# ── Redis (Upstash in prod, local in dev) ─────────────────────────────────────
redis==5.0.3

# ── File I/O ───────────────────────────────────────────────────────────────────
aiofiles==23.2.1

# ── Image & PDF Processing ─────────────────────────────────────────────────────
Pillow==10.3.0
pdf2image==1.17.0     # requires poppler-utils system package

# ── HuggingFace Inference API (HTTP client) ────────────────────────────────────
httpx==0.27.0

# ── Output Document Generators ─────────────────────────────────────────────────
python-docx==1.1.0
reportlab==4.1.0

# NOTE: No torch / transformers required!
# The model runs on HuggingFace's servers via their free Inference API.
